update test timeouts, move server to separate package
