+++ test/proxy.test.js
@@ -5,7 +5,7 @@ const { sleep } = require('../src/datatypes/util')
 
 describe('proxies client/server', function () {
   const vcount = testedVersions.length
-  this.timeout(vcount * 30 * 1000)
+  this.timeout(vcount * 7 * 60 * 1000) // upto 7 minutes per version
 
   for (const version of testedVersions) {
     it('proxies ' + version, async () => {
