+++ test/vanilla.js
@@ -1,12 +1,10 @@
-// process.env.DEBUG = 'minecraft-protocol raknet'
-const vanillaServer = require('../tools/startVanillaServer')
+process.env.DEBUG = 'minecraft-protocol raknet'
+const vanillaServer = require('minecraft-bedrock-server')
 const { Client } = require('../src/client')
 const { waitFor } = require('../src/datatypes/util')
-const { getPort } = require('./util')
+const { getPort } = require('./util/util')
 
 async function test (version) {
-  const ChunkColumn = require('bedrock-provider').chunk('bedrock_' + (version.includes('1.19') ? '1.18.30' : version)) // TODO: Fix prismarine-chunk
-
   // Start the server, wait for it to accept clients, throws on timeout
   const port = await getPort()
   const handle = await vanillaServer.startServerAndWait2(version, 1000 * 220, { 'server-port': port })
@@ -48,10 +46,6 @@ async function test (version) {
         client.queue('tick_sync', { request_time: BigInt(Date.now()), response_time: BigInt(Date.now()) })
       }, 200)
 
-      client.on('level_chunk', async packet => { // Chunk read test
-        const cc = new ChunkColumn(packet.x, packet.z)
-        await cc.networkDecodeNoCache(packet.payload, packet.sub_chunk_count)
-      })
 
       console.log('Awaiting join')
